Amend "Slightly clearer instructions in repo setup"
The description of what "yum makecache fast" does was a bit iffy.

This partially reverts commit 5a6e59065fc707838b48f1d65d7fb7bbee2058fd.

'enabled=1' is added to the repo stanza.

This removes 'fast' from the 'yum makecache' command. It was found to be
incompatible with recent versions of Fedora.

Change-Id: I669e51a792a9699fc47040620be664e7414563b6