Add RHEL 8 documentation
Updates the documentation in preparation for RHEL 8. Mostly small
changes to add the base kernel number for example (which will need
to be updated once it is officially released) as well as other
small changes. The biggest change is in the Offloadable_flows section
which adds information about setting up linux bonds.

Change-Id: I90c43e0ab3e735605b1f82ec8e930f7a3cd372f1
Signed-off-by: Louis Peens <louis.peens@netronome.com>