Update NetworkManager instructions
Current instructions are to enable NetworkManager to have persistent
config across reboots. This can interfere with TC qdisc however and
prevent offloading. Add instructions to disable NetworkManger for all
interfaces except for the PF which can still be managed by NetworkManager
to create VFs and automatically set the MTU. Also update the instructions
on how to persist VFs to use built-in settings instead of a dispatcher
script.

Change-Id: I43f5b9e826ccf73ec2683dc3378710ddb9de98c4
Signed-off-by: Louis Peens <louis.peens@netronome.com>