Update image to use external link
The Freshdesk upload API changed and does not allow for base64 encoded
images to be uploaded anymore. Update the only image in the doc to
refer to a URL on Freshdesk itself where the image is uploaded. This
has the disadvantage that the image does not show up offline and
requires the user to be able to login to view the image. Trying
this in any case as experiment to get an updated version of the
document published. This may need to be revisited in the near
future to make use of a better system to deal with images.

Change-Id: Iebbab430fb68c7ce9a79f508ebd89b7b2cc64610